% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_amr.R
\name{amr_mgen}
\alias{amr_mgen}
\title{Development of AMR}
\usage{
amr_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
either due to treatment or spontaneously(?)
}
