% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_prep.R
\name{calc_targets}
\alias{calc_targets}
\title{Calculate Network Target Statistics}
\usage{
calc_targets(
  net = nw,
  x = params,
  rel,
  count_type,
  attr_name = NULL,
  joint_attrs = c("age", "race"),
  inst_correct = FALSE,
  time_unit = "weeks"
)
}
\arguments{
\item{net}{default = "nw", the network object outputted from "generate_init_network"
(usually already in environment during workflow)}

\item{x}{default = "params", the parameter list object with parameters information for
all networks estimated from data (usually already in environment during workflow)}

\item{rel}{string, which relationship/network sub-lists to pull parameters from.}

\item{count_type}{string, which ERGM term to generate targets for. Currently can only be
in the form "edges", "nodefactor", "nodematch", "absdiff_sqrt_age", or "concurrent".}

\item{attr_name}{string, used for nodefactor and nodematch targets. If NULL, produces
targets based on the two joint_attrs specifed}
}
\description{
Calculates network targets given, at minimum, a starting network,
a list of parameters calculated from data, plus desired ERGM term and attribute.
Function currently supports calculations for edges, nodefactor, nodematch, absdiff,
and concurrent ERGM terms. Edges, nodefactor, and nodematch terms additionally require
a joint distribution (default = age and race) of parameters for calculations,
although these can be summed to output the marginal targets (e.g. nodefactor(~age))
}
