% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_startfuns.R
\name{control_mgen}
\alias{control_mgen}
\title{Epidemic Model Control Settings}
\usage{
control_mgen(
  simno = 1,
  nsims = 1,
  ncores = 1,
  nsteps = 10,
  start = 1,
  cumulative.edgelist = FALSE,
  save.nwstats = TRUE,
  nwstats.formula = "formation",
  tergmLite = TRUE,
  truncate.el.cuml = 0,
  initialize.FUN = initialize.net,
  infection.FUN = infection.net,
  recovery.FUN = recovery.net,
  prevalence.FUN = prevalence.net,
  resim_nets.FUN = resim_nets,
  summary_nets.FUN = summary_nets,
  verbose.FUN = verbose.net,
  save.network = FALSE,
  tergmLite.track.duration = FALSE,
  set.control.ergm = control.simulate.formula(MCMC.burnin = 2e+05),
  set.control.tergm = control.simulate.formula.tergm(MCMC.burnin.min = 5000),
  save.other = c("attr", "temp", "el", "net_attr"),
  verbose = FALSE,
  dat.updates = resimnet_updates_sti,
  ...
)
}
\arguments{
\item{simno}{Unique ID for the simulation run, used for file naming purposes if used
with the \code{EpiModelHPC} package.}

\item{nsims}{Number of simulations to run overall.}

\item{ncores}{Number of parallel cores per run, if parallel processing is used.}

\item{nsteps}{Number of time steps per simulation.}

\item{start}{Starting time step for simulation, with default to \code{1} to run new simulation. This
may also be set to 1 greater than the final time step of a previous simulation to resume
the simulation with different parameters.}

\item{cumulative.edgelist}{If \code{TRUE}, models tracks a cumulative edgelist (a historical list of
all partners within each network over time), useful for partner services models.}

\item{save.nwstats}{Calculate and save network statistics as defined in the
\code{simnet} modules.}

\item{nwstats.formula}{Right-hand side formula for network statistics to
monitor, with default \code{"formation"} equal to the model formula.
Supports \code{\link{multilayer}} specification.}

\item{tergmLite}{Logical indicating usage of \code{tergmLite} (non-modifiable
currently for \code{EpiModelHIV}).}

\item{truncate.el.cuml}{Number of time steps of the cumulative edgelist to retain. See help file
for \code{EpiModel::update_cumulative_edgelist} for details}

\item{initialize.FUN}{Module function to use for initialization of the epidemic model.}

\item{resim_nets.FUN}{Module function for network resimulation at each time
step.}

\item{summary_nets.FUN}{Module function for network statistic extraction at
each time step.}

\item{verbose.FUN}{Module function to print model progress to the console or
external text files.}

\item{save.network}{If \code{TRUE}, networkLite object is saved at simulation end.}

\item{tergmLite.track.duration}{If \code{TRUE}, track duration information
for models in \code{tergmLite} simulations. Supports
\code{\link{multilayer}} specification.}

\item{set.control.ergm}{Control arguments passed to \code{ergm}'s
\code{simulate_formula.network}. Supports \code{\link{multilayer}}
specification.}

\item{set.control.tergm}{Control arguments passed to \code{tergm}'s
\code{simulate_formula.network}. Supports \code{\link{multilayer}}
specification.}

\item{save.other}{A character vector of elements on the main \code{netsim_dat}
data object to save out after each simulation.}

\item{verbose}{If \code{TRUE}, print out simulation progress to the console
if in interactive mode or text files if in batch mode.}

\item{dat.updates}{See the general documentation for \code{dat.updates} at \code{\link[EpiModel:control.net]{EpiModel::control.net}}.
The standard implementation for the current MSM model is found in
\code{\link{standard_resimnet_updates}}.}

\item{...}{Additional arguments passed to the function.}

\item{aging.FUN}{Module function for aging.}

\item{departure.FUN}{Module function for general and disease-related departures.}

\item{arrival.FUN}{Module function for entries into the sexually active population.}

\item{partident.FUN}{Module function for partner identification process.}

\item{acts.FUN}{Module function to simulate the number of sexual acts within
partnerships.}

\item{condoms.FUN}{Module function to simulate condom use within acts.}

\item{prev.FUN}{Module function to calculate prevalence summary statistics.}

\item{cleanup.FUN}{Module function to tidy the \code{netsim_dat} object at the end
of each step.}
}
\value{
A list object of class \code{control.net}, which can be passed to the EpiModel function \code{\link{netsim}}.
}
\description{
Sets the controls for stochastic network models simulated with \code{\link{netsim}}.
}
