% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_prep.R
\name{generate_init_network}
\alias{generate_init_network}
\title{Generate Starting Network for ERGM Estimation}
\usage{
generate_init_network(params, deg_casual = FALSE, seed = NULL)
}
\arguments{
\item{params}{a list of parameters with a sublist called "pop" with the size of the network
and the named attributes and their distribution in the population}

\item{deg_casual}{default = FALSE, if TRUE, uses additional parameters information in "params" to set
a plausible casual degree for each node (to help fit main/long-term partnership network,
which is usually fit first in workflow)}

\item{seed}{optional, used to set a seed to maintain the same distribution of nodal attributes
every time function is called}
}
\value{
an object of class network
}
\description{
Generates empty network of specified size and adds nodal attributes.
Currently these attributes are specific to this project:
a sex variable, race, age group, age, and an adjusted age variable used to capture
reported age asymmetry in heterosexual partnerships in NSFG
(females on average slightly younger than their male partners)
}
