% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_prep.R
\name{get_nw_attr_vecs}
\alias{get_nw_attr_vecs}
\title{Extract Nodal Attribute Vectors from Network}
\usage{
get_nw_attr_vecs(nw)
}
\arguments{
\item{nw}{a network object, in this workflow, usually the network generated from "generate_init_network"}
}
\value{
A list of nodal attribute vectors
}
\description{
Outputs nodal attribute vectors as list from network object
}
