% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_startfuns.R
\name{init_mgen}
\alias{init_mgen}
\title{Epidemic Model Initial Conditions}
\usage{
init_mgen(ninfs = 5, ...)
}
\value{
A list object of class \code{init.net}, which can be passed to EpiModel function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
simulated with \code{\link{netsim}}.
}
