% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_initialize.R
\name{initialize_mgen}
\alias{initialize_mgen}
\title{Initialization Model for Mgen Model}
\usage{
initialize_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Initialize networks and starting nodal attributes
}
