% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_prep.R
\name{inst_correction}
\alias{inst_correction}
\title{Target correction for instantaneous network}
\usage{
inst_correction(targets, time_unit = NULL)
}
\arguments{
\item{targets}{a numeric vector of the ergm target statistics estimated from cumulative year data}

\item{time_unit}{either "days" or "weeks", the unit of time represented by each discrete step in simulation model}
}
\value{
A numeric vector
}
\description{
In NSFG (and other surveys), information about the frequency of one-time (instantaneous)
partnerships are reported as the cumulative number over the last 12 months. This function converts the
empirical yearly target to per-week or per-day counts.
}
