% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_prep.R
\name{matrix_symmetrical}
\alias{matrix_symmetrical}
\title{Make Empirical Mixing Matrix Symmetrical}
\usage{
matrix_symmetrical(mat)
}
\arguments{
\item{mat}{empirical mixing matrix}
}
\value{
matrix
}
\description{
Convert empirical mixing matrix to symmetrical matrix based on mean of upper/lower sections
}
