% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_transmit.R
\name{mgen_infection}
\alias{mgen_infection}
\title{Model Mgen Infection between Partners}
\usage{
mgen_infection(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Disease transmission
}
