% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixmats}
\alias{mixmats}
\title{Mixing Matricies}
\format{
The \code{mixmats} object is a
list of four mixing matrices representing:
\enumerate{
\item among main partnerships, by race
\item among main partnerships, by age group
\item among casual partnerships, by race
\item among casual partnerships, by age group
}
}
\source{
Generated using the {EpiModelSTI} R package
}
\usage{
mixmats
}
\description{
Mixing Matricies
}
\keyword{datasets}
