% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aging.R
\name{mod_aging}
\alias{mod_aging}
\title{Modules for vital dynamics}
\usage{
mod_aging(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Handles node aging, departure, and arrivals
}
