% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nw_100000}
\alias{nw_100000}
\title{Fitted Multilayer Network}
\format{
The \code{nw_100000} object is a
list of three netest objects representing main, casual, and instantaneous partnerships.
}
\source{
Generated using the {EpiModelSTI} R package
}
\usage{
nw_100000
}
\description{
Fitted Multilayer Network
}
\keyword{datasets}
