% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_startfuns.R
\name{param_mgen}
\alias{param_mgen}
\title{Epidemic Model Parameters (Mgen)}
\usage{
param_mgen(
  act_rate_main = 5.63/7,
  act_rate_casual = 1.64/7,
  act_rate_instant = 1,
  latent_period = 7,
  infectious_period = 30 * 4,
  screening_rate_females = 1/365,
  screening_rate_males = NA,
  vital = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param.net}, which can be passed to EpiModel function \code{\link{netsim}}.
}
\description{
Sets the epidemic parameters for stochastic network models
simulated with \code{\link{netsim}}
}
