% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_resim.R
\name{resim_nets_mgen}
\alias{resim_nets_mgen}
\title{Network Resimulation for Time Steps 2+}
\usage{
resim_nets_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Resimulation of dynamic networks
}
