% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafts.r, R/nw_prep.R
\name{targets}
\alias{targets}
\alias{calc_joint_nodefactor}
\alias{calc_single_attr_nodefactor}
\alias{calc_edges}
\alias{calc_absdiff}
\alias{calc_concurrent}
\alias{check_targets}
\alias{check_conditions}
\title{Target Stats Calculation Helpers}
\usage{
calc_joint_nodefactor(params, attrs, joint_attrs, joint_name, rel)

calc_single_attr_nodefactor(params, attr_name, joint_attrs, nf_joint_counts)

calc_edges(nf_joint_counts)

calc_absdiff(params, rel, count_type, edges)

calc_concurrent(params, rel, num)

check_targets(edges, final_targets, count_type, threshold = 0.01)

check_conditions(nw, params, rel, count_type, attr_name, joint_attrs)

calc_joint_nodefactor(params, attrs, joint_attrs, joint_name, rel)

calc_single_attr_nodefactor(params, attr_name, joint_attrs, nf_joint_counts)

calc_edges(nf_joint_counts)

calc_absdiff(params, rel, count_type, edges)

calc_concurrent(params, rel, num)

check_targets(edges, final_targets, count_type, threshold = 0.01)

check_conditions(nw, params, rel, count_type, attr_name, joint_attrs)
}
\arguments{
\item{params}{the parameter list object with parameters information for
all networks estimated from data}

\item{joint_attrs}{string of length 2, the two attrs used to calculate joint distribution
of estimates calculated from empirical data}

\item{joint_name}{Name of joint attribute as found in parameter input, calculated in calc_targets()}

\item{rel}{string, which relationship/network sub-lists to pull parameters from.}

\item{attr_name}{string of length 1, used for nodefactor and nodematch targets. If NULL, produces
targets based on the two joint_attrs specifed}

\item{count_type}{string, which ERGM term to generate targets for. Currently can only be
in the form "edges", "nodefactor", "nodematch", "absdiff_sqrt_age", or "concurrent".}

\item{nw}{the network object outputted from "generate_init_network"
(usually already in environment during workflow)}
}
\description{
Small helper functions used as part of calc_targets()

Small helper functions used as part of calc_targets()
}
