% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_testing.R
\name{testing_mgen}
\alias{testing_mgen}
\title{Mgen Testing and Screening}
\usage{
testing_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Regular screening and symptomatic testing
}
