% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_trackers.R
\name{trackers_mgen}
\alias{trackers_mgen}
\title{Track Metrics of Interest}
\usage{
trackers_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Track metrics of interest
}
