% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_transmit.R
\name{transmit_mgen}
\alias{transmit_mgen}
\title{Model Mgen Infection between Partners}
\usage{
transmit_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Disease transmission
}
