% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_progression.R
\name{progression_mgen}
\alias{progression_mgen}
\title{Mgen Disease Progression & Recovery}
\usage{
progression_mgen(dat, at)
}
\arguments{
\item{dat}{simulation object}

\item{at}{time step}
}
\description{
Mgen disease progression
}
